requirements.txt
Django==1.10
django-password-reset==1.0
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3.1
python-slugify==1.2.4
slugify==0.0.1
Unidecode==0.4.21
django-ckeditor==5.3.0
django-social-share==0.3.0